Create `process_transactions` module
